backend_server/requirements.txt
Django>=4.2
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.2
psycopg[binary]
python-dotenv>=1.0
jwcrypto>=1.5.6
requests>=2.31
PyJWT>=2.8
